Deleted all users in table that aren't foreign keys in `product`
Initially meant to find all unique entries and then delete everything and replace them with one of each unique entry. Except there was a bug in the middle so now all entries except for those bound by foreign key constraints in `product` are deleted.
But on the bright side all users with duplicate usernames are removed!
Code for correctly removing duplicates has been left in a block comment.