supplier table css and reroutes
changed the style of the table on the supplier.php page. I also changeed what the navbar shows based on if the url has the ?consumer_home tag after the index.php in the url. Then when the user is logged out, the ?logout tag is what tellls the session to expire so please do not change that reroute without consulting me