Password hashing and checking
Now the passwords are hashed upon being stored in the database. When a user logs in the password they entered is hashed  before being compared to the database. I also made a header on the create_account page so that if the user  successfully creates an account they are redirected to the index page