fixed #20 and cleaned up code
Alright NOW the user is properly checked for a supplier vs consumer, before it was broken and somehow half working but I have error checked it and it is properly checking the right values. I also made a new file where I dumped all the commented code from create_account into big block segments. It is within the "things we might not need" folder with "duplicate entries.php" So now that page code is cleaner. Passwords are hashed and also properly being checked which is good